Add synchronization barrier for CUDA host callbacks to allow TSAN to run
CUDA stream host callbacks are getting executed from a CUDA-managed
thread pool. Calling `cuStreamSynchronize` blocks until the
stream has finished executing all its operations including the host
callbacks.

Unfortunately thread sanitizer can't identify `cuStreamSynchronize`
as a synchronization barrier between a memory write in a host
callback and a memory read on the main thread. This is because
the CUDA libraries are not instrumented for thread sanitizer and
don't use a synchronization mechanism that could be intercepted
(they probably some variation of atomics, pthread mutexes would
have been fine).

So this change is adding an explicit synchronization barrier in
the form of a mutex to `CUDAStream`. This of course has some
unnecessary overheard but being able to run thread sanitizer is
a huge advantage so that in my opinion it's worth it.

PiperOrigin-RevId: 719201683